moved postgres into it's own module and started work on in_memory_db
