+++ persistence/src/postgres_db/mod.rs
@@ -1,0 +1,3 @@
+pub mod auth;
+pub mod postgres_db;
+pub mod users;
